third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-datachannel.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests RTCDataChannel.");

var pc = null;
var dc = null;
var data;
var array;

function dc_onclose() {
    testPassed("dc_onclose was called");
    shouldBe("dc.readyState", "'closed'");

    finishJSTest();
}

function dc_onmessage_dataview(e) {
    testPassed("dc_onmessage_dataview was called");
    data = e.data;
    shouldBe("data.byteLength", "10");
    array = new Int8Array(e.data);
    shouldBe("array[0]", "1");
    shouldBe("array[9]", "10");

    dc.onclose = dc_onclose;
    dc.close();
}

function dc_onmessage_arraybuffer(e) {
    testPassed("dc_onmessage_arraybuffer was called");
    data = e.data;
    shouldBe("data.byteLength", "2");
    array = new Int8Array(e.data);
    shouldBe("array[0]", "17");
    shouldBe("array[1]", "19");

    data = new ArrayBuffer(12);
    array = new Int8Array(data);
    array[1]=1;
    array[10]=10;

    shouldBe("data.byteLength", "12");

    shrunkView = new DataView(data, 1, 10);

    dc.onmessage = dc_onmessage_dataview;
    shouldNotThrow("dc.send(shrunkView);");
}

function dc_onmessage_string(e) {
    testPassed("dc_onmessage_string was called");
    data = e.data;
    shouldBe("data", "'xyzzy'");

    dc.binaryType = "arraybuffer";
    buffer = new ArrayBuffer(2);
    var array = new Int8Array(buffer);
    array[0] = 17;
    array[1] = 19;
    dc.onmessage = dc_onmessage_arraybuffer;
    shouldNotThrow("dc.send(buffer);");
}

function dc_onopen() {
    testPassed("dc_onopen was called");
    shouldBe("dc.readyState", "'open'");
    shouldBe("dc.label", "'label'");

    dc.onmessage = dc_onmessage_string;
    shouldNotThrow("dc.send('xyzzy');");
}

function pc_ondatachannel(e) {
    testPassed("pc_ondatachannel was called");
}

function pc_onicechange() {
    if (pc.iceConnectionState === "completed") {
        testPassed("pc is connected");
        shouldNotThrow('dc = pc.createDataChannel("label");');
        shouldBe("dc.readyState", "'connecting'");
        dc.onopen = dc_onopen;
    }
}

pc = new RTCPeerConnection();
shouldNotThrow('dc = pc.createDataChannel("label1");');
shouldBe("dc.reliable", "true");
shouldNotThrow('dc = pc.createDataChannel("label2", {});');
shouldBe("dc.reliable", "true");
shouldNotThrow('dc = pc.createDataChannel("label3", {ordered:true});');
shouldBe("dc.reliable", "true");
shouldNotThrow('dc = pc.createDataChannel("label3", {ordered:false});');
shouldBe("dc.reliable", "false");
shouldNotThrow('dc = pc.createDataChannel("label3", {maxRetransmits:0});');
shouldBe("dc.reliable", "false");
shouldNotThrow('dc = pc.createDataChannel("label3", {maxRetransmitTime:0});');
shouldBe("dc.reliable", "false");

pc = new RTCPeerConnection();
pc.oniceconnectionstatechange = pc_onicechange;
pc.ondatachannel = pc_ondatachannel;

window.jsTestIsAsync = true;
window.successfullyParsed = true;
</script>
</body>
</html>
